Extract OAuth authorization process by ActiveSupport::Concern
